refactor: extract login form from page. Prevent sidebar showing on 404 page
